Refactor MainConfig to OverrideSequence.
