+++ tutor_recon/util/constants.py
@@ -1,3 +1,5 @@
+import json
+
 import click
 import cloup
 
@@ -28,3 +30,26 @@ CONTEXT_SETTINGS = cloup.Context.settings(
 
 CONFIG_SAVE_STYLED = click.style("tutor config save", fg=EXTERNAL_COLOR)
 RECON_SAVE_STYLED = click.style("tutor recon save", fg=MAIN_COLOR)
+
+DEFAULT_OVERRIDE_SEQUENCE = json.dumps(
+    {
+        "$t": "override-sequence",
+        "overrides": [
+            {
+                "$t": "tutor",
+                "target": "config.yml",
+                "overrides": "$./tutor_config.v.json",
+            },
+            {
+                "$t": "json",
+                "target": "env/apps/openedx/config/cms.env.json",
+                "overrides": "$./openedx/cms.env.v.json",
+            },
+            {
+                "$t": "json",
+                "target": "env/apps/openedx/config/lms.env.json",
+                "overrides": "$./openedx/lms.env.v.json",
+            },
+        ],
+    }
+)
